proof of concept ipynb
built a tool in an ipynb that did the job; now I need to integrate it into the GUI. Probably as a button.

When the button is pressed, it creates the file from the current `.csv`. There might be problems with this if the `.csv` file doesn't like to be read while it is open.

Alternatively, I could create a system that writes a markdown file alongside a csv, and then when the application is closed, it creates the markdown file.